TaskFly User Service,Create,Delete,getAll und Get

+++ src/main/java/com/gocaspi/taskfly/User/UserController.java
@@ -11,7 +11,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.PathVariable;
 
 import com.google.gson.Gson;
-import com.gocaspi.taskfly.User.User;
+
+import java.util.List;
 
 
 @RestController
@@ -32,13 +33,29 @@ public class UserController {
     }
     @GetMapping("/{id}")
     public String getUser(@PathVariable String id){
-
+        if(repository.existsById(id)){
         User user = repository.findById(id).get();
         return new Gson().toJson(user);
+        }
+        else {
+            return "User Id Not Existing";
+        }
     }
     @DeleteMapping("/{id}")
-    public void deleteUser(@PathVariable String id){
+    public String deleteUser(@PathVariable String id) {
+        if(repository.existsById(id)){
+            repository.deleteById(id);
+            return "Deleted "+id+" successfully";
+        }
+        else {
+            return "User Id not Existing";
+        }
+    }
+    @GetMapping()
+    public String getAllUser(){
 
-        repository.deleteById(id);
+        List<User> users =repository.findAll();
+        return new Gson().toJson(users);
+
     }
 }
